refactor(api): decompose batch generation service
